add notification and support for MarkDown

+++ src/main/java/salmon/community/dto/QuestionDTO.java
@@ -9,13 +9,13 @@ import salmon.community.model.User;
  */
 @Data
 public class QuestionDTO {
-    private int id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
-    private long gmtCreate;
-    private long gmtModified;
-    private Integer creator;
+    private Long gmtCreate;
+    private Long gmtModified;
+    private Long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
